pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["Milos Kozak <milos.kozak@second-foundation.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tuna = "^0.5.11"
pytest-benchmark = "^4.0.0"
py-spy = "^0.3.14"
snakeviz = "^2.2.0"
pyroscope-io = "^0.8.5"
gprof2dot = "^2022.7.29"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
